[Mac] Fix crash for "settings" case of protector bubble.
If the current page was not NTP, the code to go to settings caused a
new tab to be brought up, which caused the bubble to be closed (by
base bubble controller), and then the second call to -close caused a
crash.  The duplicate calls to -close should be fine, but the message
to a freed |error_| needed to be dropped.

BUG=102681
TEST=Arrange to be on a website (like google) at startup.  Arrange to get the protector bubble up on startup, and click the settings option.  Should go to the settings, and not crash.


Review URL: http://codereview.chromium.org/8372088

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@108420 0039d316-1c4b-4281-b951-d872f2087c98