Add only types with empty progress marker to config requests. This fixes a bug where in case of a newly enabled datatype we would end up adding all the datatypes to the config request.
With this change Nigori would not be part of the requested types unless it has progress marker token to be empty. So we should not get migration error on config requests at all. In case we do it would be a unrecoverable error which is fine.


BUG=102557
TEST=


Review URL: http://codereview.chromium.org/8439057

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@108620 0039d316-1c4b-4281-b951-d872f2087c98