WebUI test framework: fix Mock4JS verification, allow chrome.send passthrough.
1) Add test for Bug 99970 Comment 6 & fix
2) Allow chrome.send to be mocked outside of preLoad call.
3) add chrome.originalSend to allow mocked messages to pass through to C++.

Also to address need for .cc file to be correct for clang, and prevent ugliness in js with a lot of GEN lines, allow the header and .cc to not have a suffix, and the generated file to end in -gen.cc.  While doing this, addressed crbug.com/102794 to put in <(INTERMEDIATE_DIR) preventing collisions across rules.

R=flackr@chromium.org,wyck@chromium.org
BUG=103740,99970,102794
TEST=browser_tests --gtest_filter=Mock4JS*.* AND ChromeSend*.*

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=109678

Review URL: http://codereview.chromium.org/8515016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@109711 0039d316-1c4b-4281-b951-d872f2087c98