Add a new field trial for Instant.
The new field trial (SUGGEST) is almost the same as HIDDEN, except that the
Instant suggested text is allowed to be autocompleted into the omnibox. Since
the hidden tab contains results for the full query, it's more likely to be
relevant when the user presses <Enter>. From the user perspective, the omnibox
autocompletion is similar to how it behaves for past searches.

BUG=none
TEST=With --instant-field-trial=suggest, type into the omnibox, and observe
that it gets inline autocompleted, even though there's no Instant preview.


Review URL: http://codereview.chromium.org/8588003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110463 0039d316-1c4b-4281-b951-d872f2087c98