Remove nord=1 from the Instant template for Google.
This was originally added to prevent the homepage from redirecting to
iGoogle, which doesn't support Instant. However, this param also
prevents redirecting to SSL, which redirect we actually want.

With this change, iGoogle users will see no Instant previews (so they'll
have to press <Enter> to search), but it is being fixed transparently on
the server side, so that Chrome Instant homepage requests will never go
to iGoogle. That fix should be live by the end of the next week.

BUG=105669
TEST=Sign in to Google. With Instant enabled, do a query. Observe that you are on SSL (https).

Review URL: http://codereview.chromium.org/8733003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@112088 0039d316-1c4b-4281-b951-d872f2087c98