Add more CHECK in GPU and NPAPI process on invalid file descriptors
Goes with r112647, where we added CHECKs for invalid file descriptors.

Regarding the GPU process, we already have a crash in GpuChannelManager::OnEstablishChannel() function due to invalid renderer fd.  But it's not clear whether this is due to channel reuse or failure in creating channel in server mode.  Talked to apatrick@ and the reuse one should never happen in normal case.  Adding check on that case so we can hopefully rule out that case.

Regarding the NPAPI process, we already have a lot of crashes due to invalid file descriptors in the rederer.  In NPAPI process, add a CHECK right after the channel is created in server mode to see where this invalid fd originates from.

BUG=none
TEST=passed unit tests


Review URL: http://codereview.chromium.org/8877001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@113767 0039d316-1c4b-4281-b951-d872f2087c98