Make hosted apps launched via chrome.tabs.create with an URL redirect have the right process type.
(This affected apps like Google Docs when launched from the bookmark manager)

The check for UseWebUIForURL in RenderViewHostManager::GetSiteInstanceForEntry
was not enough, since it only returns true for URLs that need extension
bindings, which is not true for hosted apps.

This only affected chrome.tabs.create, and not regular <a target="_blank"> links
since those are first created in the opening tab's process, thus the current
site instance is the launching extension's, so we would consider them a
WebUI -> non-WebUI transition, which needs a process swap.

BUG=99349
R=creis@chromium.org


Review URL: http://codereview.chromium.org/8773050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@114733 0039d316-1c4b-4281-b951-d872f2087c98