Move srpcgen running out of scons and into its own script
The logic for running srpcgen.py was littered in various scons bits,
some in the chromium repository and some in the native_client repository.
Consolidate all of it into a standalone Python script and remove the old
stuff.  The scons-driven diff test is now just a simple invocation of this
script with --diff_mode.

BUG= none
TEST= cd native_client; ./scons srpcdif

R=noelallen@chromium.org,ncbray@google.com

Review URL: http://codereview.chromium.org/8983033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@116521 0039d316-1c4b-4281-b951-d872f2087c98